Aktualizacja projektu
