+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@react-google-maps/api": "^2.20.6",
     "@supabase/supabase-js": "^2.49.1",
+    "lottie-react": "^2.4.1",
     "lucide-react": "^0.483.0",
     "next": "15.2.3",
     "react": "^19.0.0",
