+++ src/app/globals.css
@@ -10,6 +10,7 @@ body {
   .font-permanent-marker {
     font-family: var(--font-permanent-marker);
   }
+
   @keyframes slideInRight {
     from {
       opacity: 0;
@@ -20,7 +21,7 @@ body {
       transform: translateX(0);
     }
   }
-  
+
   @keyframes fadeDown {
     from {
       opacity: 0;
@@ -31,15 +32,26 @@ body {
       transform: translateY(0);
     }
   }
-  
+
+  @keyframes fadeIn {
+    from {
+      opacity: 0;
+      transform: scale(0.95);
+    }
+    to {
+      opacity: 1;
+      transform: scale(1);
+    }
+  }
+
   .animate-slide-in-right {
     animation: slideInRight 1s ease-out forwards;
   }
-  
+
   .animate-fade-down {
     animation: fadeDown 0.8s ease-out forwards;
   }
-  
+
   .animate-delay-1 {
     animation-delay: 0.2s;
   }
@@ -52,5 +64,8 @@ body {
   .animate-delay-4 {
     animation-delay: 0.8s;
   }
-  
+
+  .animate-fadeIn {
+    animation: fadeIn 0.5s ease-in-out;
+  }
 }
